Move expressions to metabase-lib
